MHER-37: Add a test rapidpro server. Add a server unavailable test
CR suggestions
